+++ tests/test_time_series_graph.py
@@ -720,6 +720,26 @@ class TestTimeSeriesCausalGraph(unittest.TestCase):
         # deep equality
         self.assertTrue(graph.__eq__(summary_graph, True))
 
+        # Try with floating nodes.
+        # X[t], Y[t-1]->Z[t], the summary graph should be X, Y->Z.
+        graph = CausalGraph()
+        graph.add_node('x')
+        graph.add_edge('y', 'z')
+        tsgraph = TimeSeriesCausalGraph()
+        tsgraph.add_node('x')
+        tsgraph.add_edge('y lag(n=1)', 'z')
+
+        summary_graph = tsgraph.get_summary_graph()
+
+        self.assertEqual(summary_graph, graph)
+        self.assertTrue(graph.__eq__(summary_graph, True))  # deep equality
+
+        # Make sure extra floating of same var name is fine.
+        tsgraph.add_node('x lag(n=1)')
+        summary_graph = tsgraph.get_summary_graph()
+        self.assertEqual(summary_graph, graph)
+        self.assertTrue(graph.__eq__(summary_graph, True))  # deep equality
+
     def test_variable_names(self):
         variables = self.tsdag.variables
         self.assertEqual(variables, ['X1', 'X2', 'X3'])